+++ src/Backend/Items/spItems.ts
@@ -33,7 +33,8 @@ export interface SPItem {
 
   toCompletionItem(
     lastFunc?: MethodItem | FunctionItem | undefined,
-    lastESOrMM?: MethodMapItem | EnumStructItem | undefined
+    lastESOrMM?: MethodMapItem | EnumStructItem | undefined,
+    override?: boolean
   ): CompletionItem | undefined;
   toDefinitionItem(): LocationLink | undefined;
   toReferenceItem?(): Location[];